backend/requirements.txt
anyio==4.9.0
click==8.1.8
fastapi==0.95.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
pillow==11.2.1
pydantic==1.10.21
python-multipart==0.0.20
setuptools==78.1.0
sniffio==1.3.1
starlette==0.26.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.2
uvicorn==0.21.0
